api/requirements.txt
flask
flask_restful
mysql-connector-python
Flask-CORS
PyJWT